refactor(tests): move test framework to library/ folder
The split is not 100% complete yet, but it's a good start. Eventually I
want to move the non Neovim specific parts of the test framework to
their own library and publish that separately.
