+++ integration-tests/client/__global.ts
@@ -1,10 +1,13 @@
 import type {
   StartNeovimArguments,
   TestDirectory,
-} from "server/application/neovim/environment/testEnvironmentTypes"
+} from "../library/server/application/neovim/environment/testEnvironmentTypes.ts"
 
+// This defines a way for the test runner to start Neovim
 declare global {
   interface Window {
     startNeovim(startArguments?: StartNeovimArguments): Promise<TestDirectory>
   }
 }
+
+export {}
