+++ integration-tests/client/style.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: "DejaVuSansMNerdFontMono";
-  src: url("/DejaVuSansMNerdFontMono-Regular.ttf");
+  src: url("../library/public/DejaVuSansMNerdFontMono-Regular.ttf");
   font-weight: normal;
   font-style: normal;
 }
