+++ integration-tests/cypress/e2e/using-ya-to-read-events/startNeovimWithYa.ts
@@ -1,7 +1,7 @@
 import type {
   StartNeovimArguments,
   TestDirectory,
-} from "server/application/neovim/environment/testEnvironmentTypes"
+} from "library/server/application/neovim/environment/testEnvironmentTypes"
 
 /** NOTE: always uses the `modify_yazi_config_to_use_ya_as_event_reader.lua` as
  * that is implied by the name of the function.
