+++ integration-tests/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/client/websocket-client.ts"></script>
+    <script type="module" src="./client/neovim-client.ts"></script>
   </body>
 </html>
