+++ integration-tests/library/client/validateMouseEvent.ts
