+++ integration-tests/library/client/vite-env.d.ts
