+++ integration-tests/library/public/DejaVuSansMNerdFontMono-Regular.ttf
