+++ integration-tests/library/public/yazi-logo.png
