+++ integration-tests/package.json
@@ -9,7 +9,7 @@
     "cy:run:ci": "concurrently --success command-cypress --kill-others --names 'client,server,cypress' --prefix-colors 'blue,green,yellow' 'npm run dev:client' 'npm run dev:server' 'wait-on --timeout 60000 http://127.0.0.1:5173 && npx cypress run'",
     "dev": "concurrently --kill-others --names 'client,server,cypress' --prefix-colors 'blue,green,yellow' 'npm run dev:client' 'npm run dev:server' 'npm run cy:open'",
     "dev:client": "vite",
-    "dev:server": "nodemon --watch ./server --ext '*' --exec 'npx tsx ./server/server.ts'",
+    "dev:server": "nodemon --watch ./library/server --ext '*' --exec 'npx tsx ./library/server/server.ts'",
     "eslint": "eslint --max-warnings=0 .",
     "preview": "vite preview"
   },
