+++ integration-tests/tsconfig.json
@@ -24,5 +24,5 @@
     "noFallthroughCasesInSwitch": true
   },
 
-  "include": ["./client", "./server", "./cypress", "./eslint.config.mjs"]
+  "include": ["./library", "./client", "./cypress", "./eslint.config.mjs"]
 }
